requirements.txt
appdirs==1.4.4
attrs==19.3.0
backcall==0.2.0
bleach==3.1.5
certifi==2019.11.28
cffi==1.14.1
Click==7.0
colorama==0.4.3
cryptography==3.0
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.1
dnspython==1.16.0
dominate==2.4.0
entrypoints==0.3
eventlet==0.25.1
filelock==3.0.12
Flask==1.1.1
Flask-Bootstrap==3.3.7.1
Flask-MQTT==1.0.5
Flask-SocketIO==3.3.2
greenlet==0.4.15
ipykernel==5.3.2
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
jedi==0.17.1
Jinja2==2.11.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.5
jupyter-console==6.1.0
jupyter-core==4.6.3
MarkupSafe==1.1.1
mistune==0.8.4
monotonic==1.5
nbconvert==5.6.1
nbformat==5.0.7
notebook==6.0.3
packaging==20.4
paho-mqtt==1.5.0
pandocfilters==1.4.2
parso==0.7.0
passlib==1.7.2
pickleshare==0.7.5
prometheus-client==0.8.0
prompt-toolkit==3.0.5
pycparser==2.20
Pygments==2.6.1
PyMySQL==0.10.0
pyparsing==2.4.7
pyrsistent==0.16.0
python-dateutil==2.8.1
python-engineio==3.11.2
python-socketio==4.4.0
pywin32==228
pywinpty==0.5.7
pyzmq==19.0.1
qtconsole==4.7.5
QtPy==1.9.0
Send2Trash==1.5.0
six==1.14.0
terminado==0.8.3
testpath==0.4.4
tornado==6.0.4
traitlets==4.3.3
typing==3.7.4.1
virtualenv==20.0.27
visitor==0.1.3
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.0
widgetsnbextension==3.5.1
